requirements.txt
flask
tensorflow
keras
librosa
numpy
scipy
soundfile
pydub
pymongo
python-dotenv
scikit-learn
